Fixed major bug
I was unaware that there were asynchronous calls in the user loop within
the updateLeaderboard function. This caused inconsistencies when
updating the database which messed up the pull request list for at least
one user.

Fixed it by first making helper functions to make the function more
manageable, then used a step function to replace the user loop. To make
sure there wasn't a call stack overflow, I called setTimeout. This seems
to work, if it doesn't scale well for some reason I'll try to work with
the asynchrony, or just run the updateLeaderboard code on a seperate
server.